+++ features/support/datamapper.rb
@@ -1,0 +1,5 @@
+require 'dm-core'
+require 'carrierwave/mount'
+require 'carrierwave/orm/datamapper'
+
+DataMapper.setup(:default, 'sqlite3::memory:')