+++ features/support/env.rb
@@ -2,6 +2,8 @@ $TESTING=true
 $:.push File.join(File.dirname(__FILE__), '..', 'lib')
 
 require 'rubygems'
+require File.join(File.dirname(__FILE__), 'activerecord')
+require File.join(File.dirname(__FILE__), 'datamapper')
 
 if ENV["AS"]
   puts "--> using ActiveSupport"
@@ -20,7 +22,7 @@ require 'carrierwave'
 alias :running :lambda
 
 def file_path( *paths )
-  File.expand_path(File.join(File.dirname(__FILE__), *paths))
+  File.expand_path(File.join(File.dirname(__FILE__), '..', *paths))
 end
 
 CarrierWave.config[:public] = file_path('public')