+++ lib/carrierwave/processing/rmagick.rb
@@ -63,6 +63,37 @@ module CarrierWave
   #
   module RMagick
 
+    def self.included(base)
+      super
+      base.extend(ClassMethods)
+    end
+
+    module ClassMethods
+      def convert(format)
+        process :resize_to_limit => format
+      end
+
+      def resize_to_limit(width, height)
+        process :resize_to_limit => [width, height]
+      end
+
+      def resize_to_fit(width, height)
+        process :resize_to_fit => [width, height]
+      end
+
+      def resize_to_fill(width, height)
+        process :resize_to_fill => [width, height]
+      end
+
+      def resize_and_pad(width, height)
+        process :resize_to_fit => [width, height]
+      end
+
+      def resize_and_pad(width, height, background=:transparent, gravity=::Magick::CenterGravity)
+        process :resize_and_pad => [width, height, background, gravity]
+      end
+    end
+
     ##
     # Changes the image encoding format to the given format
     #
@@ -109,7 +140,7 @@ module CarrierWave
         new_img = img.change_geometry(geometry) do |new_width, new_height|
           img.resize(new_width, new_height)
         end
-        img.destroy!
+        destroy_image(img)
         new_img = yield(new_img) if block_given?
         new_img
       end
@@ -194,9 +225,9 @@ module CarrierWave
         else
           filled = new_img.color_floodfill(1, 1, ::Magick::Pixel.from_color(background))
         end
-        new_img.destroy!
+        destroy_image(new_img)
         filled.composite!(img, gravity, ::Magick::OverCompositeOp)
-        img.destroy!
+        destroy_image(img)
         filled = yield(filled) if block_given?
         filled
       end
@@ -231,14 +262,20 @@ module CarrierWave
           list << yield( frame )
         end
         list.write(current_path)
-        list.destroy!
+        destroy_image(list)
       else
         frame = image.first
         yield( frame ).write(current_path)
-        frame.destroy!
+        destroy_image(frame)
       end
     rescue ::Magick::ImageMagickError => e
       raise CarrierWave::ProcessingError.new("Failed to manipulate with rmagick, maybe it is not an image? Original Error: #{e}")
+    end
+
+  private
+  
+    def destroy_image(image)
+      image.destroy! if image.respond_to?(:destroy!)
     end
 
   end # RMagick