+++ lib/carrierwave/uploader/callbacks.rb
@@ -7,31 +7,31 @@ module CarrierWave
       end
 
       def with_callbacks(kind, *args)
-        self.class._before_callback(kind).each { |callback| self.send(callback, *args) }
+        self.class._before_callbacks_for(kind).each { |callback| self.send(callback, *args) }
         yield
-        self.class._after_callback(kind).each { |callback| self.send(callback, *args) }
+        self.class._after_callbacks_for(kind).each { |callback| self.send(callback, *args) }
       end
 
       module ClassMethods
         
-        def _before_callback(kind) #:nodoc:
+        def _before_callbacks_for(kind) #:nodoc:
           self._before_callbacks ||= {}
           self._before_callbacks[kind] ||= []
           self._before_callbacks[kind]
         end
 
-        def _after_callback(kind) #:nodoc:
+        def _after_callbacks_for(kind) #:nodoc:
           self._after_callbacks ||= {}
           self._after_callbacks[kind] ||= []
           self._after_callbacks[kind]
         end
 
         def before(kind, callback)
-          _before_callback(kind) << callback
+          _before_callbacks_for(kind) << callback
         end
 
         def after(kind, callback)
-          _after_callback(kind) << callback
+          _after_callbacks_for(kind) << callback
         end
       end # ClassMethods